Worked on modify and jotted down ideas.

+++ CharacterCreator.py
@@ -1,6 +1,7 @@
 #character creator class
 
-
+import os
+import shutil
 #implements the character creator class
 
 
@@ -8,10 +9,25 @@
 #	create - creates character
 #	newCharacteroptions - after new character created
 #	modify - modify character file
+#	changecharactersname - change the name of the character thus changing the file name
 #	deletecharacter - delete a character
 #	listallCharacters - lists all character files
 
 #class CharacterCreator:
+#Remember to implement a way so it reads from directory of program
+#Make a folder for characters?
+
+#Maybe first do a command to find the base file
+#Start search at home
+#Then make current dir the dir it is found in 
+
+#changechactername - os.rename
+
+
+# print os.getcwd() - current work directory
+# os.system() - use a shell command move to correct dir?
+
+
 
 #create character
 def create():
@@ -27,7 +43,9 @@ def create():
 		
 		#check if there already is a file with that name?
 		#copies base file and renames with character name
-		#could use UNIX commands here (find, copy, rename)
+		#could maybe use UNIX commands here (find, copy, rename)
+		open(filename, 'a').close()
+		print shutil.copy('Characterbase.txt', filename)	
 		
 	else:
 		create( )
@@ -37,34 +55,43 @@ def modify():
 	firstname = raw_input('Characters first name:')
 	filename = firstname + ".txt";
 	
-	#finds file (use UNIX find)
-	#if file not found, print message saying say so
+	#searches for file
+	if os.path.exists(filename):
 	
-	#opens file if file exists
-	charfile = open(filename, 'w') 
+		#opens file if file exists
+		charfile = open(filename, 'w') 
 	
-	#reads off lines with [#]
-	#lets user type in or skip
-	#updates as it goes
+		#reads off lines with [#]
+		#displays what is already written too - this needs to be editable
+		#lets user type in or skip
+		#updates as it goes
 	
-	#once end of file reached, say so 
-	#close file
+		#once end of file reached, say so
+		print 'End of the file reached'
+		
+		#close file
+		
+	else:
+		print 'Error: File not found', filename
 	
 def deletecharacter():
 	#ask for character name
 	firstname = raw_input('Characters first name:')
 	filename = firstname + ".txt";
 	
-	#finds file (use UNIX find)
-	#if file not found, print message saying say so
+	#searches for file
+	if os.path.exists(filename):
 	
-	#asks if user is sure
-	print 'Press Y if you would like to delete the character', firstname
-	correct = raw_input()
+		#asks if user is sure
+		print 'Press Y if you would like to delete the character', firstname
+		correct = raw_input()
 
-	#If Y pressed, delete file (UNIX delete)
-	if correct == "Y":
+		#If Y pressed, delete file (UNIX delete)
+		if correct == "Y":
 		#delete
+		
+	else:
+		print 'Error: File not found', filename
 
 	
 def listallCharacters():
